Refactor #11 네이밍 리팩토링
