+++ app/src/main/java/com/example/messenger/repository/message/MessageRepository.kt
@@ -1,6 +1,6 @@
 package com.example.messenger.repository.message
 
-import com.example.messenger.repository.model.Message
+import com.example.messenger.repository.model.chat.Message
 import io.reactivex.Completable
 import io.reactivex.Flowable
 import io.reactivex.Single
