+++ app/src/main/java/com/example/messenger/repository/message/MessageRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.example.messenger.repository.message
 
 import com.example.messenger.database.message.MessageDatabase
 import com.example.messenger.network.SocketHelper
-import com.example.messenger.repository.model.Message
+import com.example.messenger.repository.model.chat.Message
 import io.reactivex.Completable
 import io.reactivex.Flowable
 import io.reactivex.Single
