+++ app/src/main/java/com/example/messenger/ui/chatroom/ChatRoomActivityHelper.kt
@@ -1,15 +1,15 @@
-package com.example.messenger.ui.chattingroom
+package com.example.messenger.ui.chatroom
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import com.example.messenger.base.BaseHelper
-import com.example.messenger.ui.chattingroom.adapter.MessageRecyclerViewListener
+import com.example.messenger.ui.chatroom.adapter.MessageRecyclerViewListener
 import com.example.messenger.usecase.LoadMessagesUseCase
 
 /**
  * @author bsgreentea
  */
-class ChattingRoomActivityHelper(
+class ChatRoomActivityHelper(
     private val loadMessagesUseCase: LoadMessagesUseCase,
     private val roomId: String
 ) : BaseHelper {
