+++ app/src/main/java/com/example/messenger/ui/chatroom/ChatRoomViewModelFactory.kt
@@ -1,4 +1,4 @@
-package com.example.messenger.ui.chattingroom
+package com.example.messenger.ui.chatroom
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -7,13 +7,13 @@ import com.example.messenger.repository.message.MessageRepositoryImpl
 /**
  * @author bsgreentea
  */
-class ChattingRoomViewModelFactory(
+class ChatRoomViewModelFactory(
     val repository: MessageRepositoryImpl,
     val roomId: String
 ) : ViewModelProvider.NewInstanceFactory() {
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return ChattingRoomViewModel(repository, roomId) as T
+        return ChatRoomViewModel(repository, roomId) as T
     }
 }
