+++ app/src/main/java/com/example/messenger/ui/chatroom/ChatRoomViewModelInjector.kt
@@ -1,15 +1,15 @@
-package com.example.messenger.ui.chattingroom
+package com.example.messenger.ui.chatroom
 
 import com.example.messenger.repository.message.MessageRepositoryInjector
 
 /**
  * @author bsgreentea
  */
-object ChattingRoomViewModelInjector {
+object ChatRoomViewModelInjector {
     fun provideSampleViewModelFactory(
         roomId: String
-    ): ChattingRoomViewModelFactory {
+    ): ChatRoomViewModelFactory {
         val repository = MessageRepositoryInjector.getMessageRepositoryImpl()
-        return ChattingRoomViewModelFactory(repository, roomId)
+        return ChatRoomViewModelFactory(repository, roomId)
     }
 }
