+++ app/src/main/java/com/example/messenger/ui/chatroom/adapter/MessageReceivedItemViewModel.kt
@@ -1,0 +1,12 @@
+package com.example.messenger.ui.chatroom.adapter
+
+import com.example.messenger.repository.model.chat.Message
+
+/**
+ * @author bsgreentea
+ */
+class MessageReceivedItemViewModel(message: Message) {
+    val senderName = message.senderUid
+    val messageContent = message.message
+    val messageTime = message.timestamp
+}
