+++ app/src/main/java/com/example/messenger/ui/chatroom/adapter/MessageSentItemViewModel.kt
@@ -1,0 +1,11 @@
+package com.example.messenger.ui.chatroom.adapter
+
+import com.example.messenger.repository.model.chat.Message
+
+/**
+ * @author bsgreentea
+ */
+class MessageSentItemViewModel(message: Message) {
+    val messageContent = message.message
+    val messageTime = message.timestamp
+}
