--- app/src/main/java/com/example/messenger/ui/chattingroom/adapter/MessageSentItemViewModel.kt
@@ -1,11 +1,0 @@
-package com.example.messenger.ui.chattingroom.adapter
-
-import com.example.messenger.repository.model.Message
-
-/**
- * @author bsgreentea
- */
-class MessageSentItemViewModel(message: Message) {
-    val messageContent = message.messageContent
-    val messageTime = message.messageTime
-}
