+++ app/src/main/java/com/example/messenger/usecase/LoadMessagesUseCase.kt
@@ -2,7 +2,7 @@ package com.example.messenger.usecase
 
 import com.example.messenger.event.ChattingRoomEvent
 import com.example.messenger.repository.message.MessageRepositoryImpl
-import com.example.messenger.repository.model.Message
+import com.example.messenger.repository.model.chat.Message
 import io.reactivex.disposables.CompositeDisposable
 
 /**
