+++ app/src/main/java/com/example/messenger/usecase/ReceiveMessageUseCase.kt
@@ -13,7 +13,7 @@ package com.example.messenger.usecase
 
 import android.util.Log
 import com.example.messenger.repository.message.MessageRepositoryImpl
-import com.example.messenger.repository.model.Message
+import com.example.messenger.repository.model.chat.Message
 import com.example.messenger.tools.convertJsonStringToMessage
 import io.reactivex.disposables.CompositeDisposable
